Response entity

+++ src/main/java/net/harsh/journalApp/Controller/JournalEntryControllerv2.java
@@ -4,10 +4,13 @@ import net.harsh.journalApp.entity.JournalEntryv2;
 import net.harsh.journalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/journal")
@@ -17,36 +20,61 @@ public class JournalEntryControllerv2 {
     private JournalEntryService journalEntryService;
 
     @GetMapping
-    public List<JournalEntryv2> getAll() {
-        return journalEntryService.getAll();
+    public ResponseEntity<List<JournalEntryv2>> getAll() {
+        List<JournalEntryv2> allEntries = journalEntryService.getAll();
+        if (allEntries != null && !allEntries.isEmpty()) {
+            return new ResponseEntity<>(allEntries, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping
-    public JournalEntryv2 createEntry(@RequestBody JournalEntryv2 journalEntry) {
-        journalEntry.setDate(LocalDateTime.now());
-        journalEntryService.saveEntry(journalEntry);
-        return journalEntry;
+    public ResponseEntity<JournalEntryv2> createEntry(@RequestBody JournalEntryv2 journalEntry) {
+        try {
+            journalEntry.setDate(LocalDateTime.now());
+            journalEntryService.saveEntry(journalEntry);
+            return new ResponseEntity<>(journalEntry, HttpStatus.CREATED);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/{journalId}")
-    public JournalEntryv2 getJournalById(@PathVariable ObjectId journalId) {
-        return journalEntryService.findById(journalId).orElse(null);
+    public ResponseEntity<JournalEntryv2> getJournalById(@PathVariable ObjectId journalId) {
+        Optional<JournalEntryv2> journalEntry = journalEntryService.findById(journalId);
+//        if (journalEntry.isPresent()) {
+//            return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
+//        }
+//        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        // better practice
+        return journalEntry.map(entry -> new ResponseEntity<>(entry, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PutMapping("/{journalId}")
-    public JournalEntryv2 updateEntry(@PathVariable ObjectId journalId, @RequestBody JournalEntryv2 newEntry) {
-        JournalEntryv2 journalEntry = journalEntryService.findById(journalId).orElse(null);
-        if (journalEntry != null) {
-            journalEntry.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().isEmpty() ? newEntry.getTitle() : journalEntry.getTitle());
-            journalEntry.setContent(newEntry.getContent() != null && !newEntry.getContent().isEmpty() ? newEntry.getContent() : journalEntry.getContent());
+    public ResponseEntity<JournalEntryv2> updateEntry(@PathVariable ObjectId journalId, @RequestBody JournalEntryv2 newEntry) {
+        Optional<JournalEntryv2> optional = journalEntryService.findById(journalId);
+        if (!optional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        JournalEntryv2 journalEntry = optional.get();
+        if (newEntry.getTitle() != null && !newEntry.getTitle().isEmpty()) {
+            journalEntry.setTitle(newEntry.getTitle());
+        }
+        if (newEntry.getContent() != null && !newEntry.getContent().isEmpty()) {
+            journalEntry.setContent(newEntry.getContent());
         }
         journalEntryService.saveEntry(journalEntry);
-        return journalEntry;
+        return new ResponseEntity<>(journalEntry, HttpStatus.OK);
     }
 
     @DeleteMapping("/{journalId}")
-    public boolean deleteJournalEntry(@PathVariable ObjectId journalId) {
+    public ResponseEntity<?> deleteJournalEntry(@PathVariable ObjectId journalId) {
+        if (!journalEntryService.findById(journalId).isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
         journalEntryService.deleteById(journalId);
-        return true;
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
